OCR Added to main.py
