+++ backend/main.py
@@ -1,11 +1,14 @@
-from fastapi import FastAPI, HTTPException, Body
 import os
-from dotenv import load_dotenv
-from models import User # import the user model defined by us
-from motor.motor_asyncio import AsyncIOMotorClient
 from contextlib import asynccontextmanager
-from typing import List 
-from pydantic import BaseModel # Most widely used data validation library for python
+from typing import List
+
+from dotenv import load_dotenv
+from fastapi import FastAPI, HTTPException, Body, File, UploadFile
+from motor.motor_asyncio import AsyncIOMotorClient
+from pydantic import BaseModel  # Most widely used data validation library for python
+
+from OCR import scan  # imports user defined scan func for reading bills
+from models import User  # import the user model defined by us
 
 load_dotenv()
 
@@ -20,20 +23,22 @@ async def lifespan(app: FastAPI):
     # Close the database connection
     await shutdown_db_client(app)
 
+
 # method for start the MongoDb Connection
 async def startup_db_client(app):
     app.mongodb_client = AsyncIOMotorClient(os.getenv("MONGODB_URL"))
     app.mongodb = app.mongodb_client.get_database("Database")
     print("MongoDB connected.")
 
+
 # method to close the database connection
 async def shutdown_db_client(app):
     app.mongodb_client.close()
     print("Database disconnected.")
 
+
 # creating a server with python FastAPI
 app = FastAPI(lifespan=lifespan)
-
 
 origins = ["*"]
 
@@ -45,13 +50,16 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+
 @app.get("/")
 async def read_root():  # function that is binded with the endpoint
     return {"Hello": "World"}
 
+
 class loginInfo(BaseModel):
-    email_address:str
+    email_address: str
     password: str
+
 
 @app.post("/api/v1/auth/login", response_model=User)
 async def login_user(userInfo: loginInfo = Body(...)):
@@ -81,6 +89,7 @@ async def insert_user(user: User):
     inserted_user = await app.mongodb["users"].find_one({"_id": result.inserted_id})
     return inserted_user
 
+
 # R <=== Read
 # Read all users
 @app.get("/api/v1/read-all-users", response_model=List[User])
@@ -88,12 +97,25 @@ async def read_users():
     users = await app.mongodb["users"].find().to_list(None)
     return users
 
+
 # Read one user by email_address
 @app.get("/api/v1/read-user/{email_address}", response_model=User)
 async def read_user_by_email(email_address: str):
     user = await app.mongodb["users"].find_one({"email_address": email_address})
-    
+
     if user is None:
         raise HTTPException(status_code=404, detail="User not found")
     return user
 
+
+# ocr
+@app.post("/upload/")
+async def upload_image(image: UploadFile = File(...)):
+    image.filename = "/Users/rishitbaitule/PycharmProjects/Expense-Tracker/images/img.png"
+    contents = await image.read()
+
+    with open(f"{image.filename}", "wb") as f:
+        f.write(contents)
+        cost = scan(image.filename)
+    os.remove(image.filename)
+    return {"filename": image.filename, "cost": cost}
